Near working inflate process

+++ lib/App/KSP_CKAN/Crawler.pm
@@ -4,6 +4,7 @@ use strict;
 use warnings;
 use v5.010;
 use File::Find::Rule;
+use File::Find::Age;
 use File::Copy qw(copy);
 use File::Temp qw(tempdir);
 use File::Path qw(remove_tree mkpath);
@@ -13,6 +14,8 @@ use App::KSP_CKAN::Tools::IA;
 use App::KSP_CKAN::Metadata::Ckan;
 use App::KSP_CKAN::Crawler::Schema;
 use App::KSP_CKAN::Crawler::InflateNetKAN;
+use App::KSP_CKAN::Crawler::MirrorCKAN;
+use List::MoreUtils qw(first_index any);
 use Method::Signatures 20140224;
 use Carp qw(croak);
 use AnyEvent::ForkManager;
@@ -47,7 +50,6 @@ has '_schema'         => ( is => 'rw', lazy => 1, builder => 1 );
 has '_CKAN_meta'      => ( is => 'ro', lazy => 1, builder => 1 );
 has '_CKAN_meta_path' => ( is => 'ro', lazy => 1, builder => 1 );
 has '_output'         => ( is => 'ro', lazy => 1, builder => 1 );
-has '_inflator'       => ( is => 'ro', lazy => 1, builder => 1 );
 has '_ia'             => ( is => 'ro', lazy => 1, builder => 1 );
 has '_inflating'      => ( is => 'rw', default => sub { 0 } );
 has '_mirror_check'   => ( is => 'rw', default => sub { 0 } );
@@ -66,12 +68,6 @@ method _build__CKAN_meta {
     remote  => $self->config->CKAN_meta,
     local   => $self->config->working,
     clean   => 1,
-  );
-}
-
-method _build__inflator {
-  return App::KSP_CKAN::Crawler::InflateNetKAN->new(
-    config  => $self->config,
   );
 }
 
@@ -140,7 +136,7 @@ method check_random_mirrored($number = 20) {
   $self->debug("Checking random mirrored ckans");
   foreach my $row (@rows) {
     my $ckan = App::KSP_CKAN::Metadata::Ckan->new( file => $path."/".$row->identifier."/".$row->file );
-    $self->debug("Checking ".$ckan->mirror_item." - SHA1: ".$ckan->download_sha1) if $ckan->can_mirror;
+    $self->info("Checking mirror for: ".$ckan->mirror_item." - SHA1: ".$ckan->download_sha1) if $ckan->can_mirror;
     $row->update( {
       last_checked => \'NOW()',
       mirrored    => $self->_ia->ckan_mirrored( ckan => $ckan),
@@ -149,15 +145,20 @@ method check_random_mirrored($number = 20) {
   }
 }
 
-method inflate_random($number = 2) {
-  my @rows  = $self->_schema->resultset('CKAN_meta')->rand($number)->search({ deleted => 0, download_sha1 => 0 });
+method _check_output($path, $file) {
+  my $newfile = File::Find::Age->in($path)->[-1]->{file};
+  return basename($newfile);
+}
+
+# TODO: Oh gosh this is unwieldy, could do with a refactor.
+method inflate_random($number = 1) {
+  my @rows = $self->_schema->resultset('CKAN_meta')->rand($number)->search({ deleted => 0, download_sha1 => 0 });
   my $path = $self->_CKAN_meta_path;
-  $self->_CKAN_meta->pull;
   
   my $tmp = File::Temp::tempdir();
   my @files;
   $self->debug("Inflating random ckans");
-  
+ 
   foreach my $row (@rows) {
     $self->debug("Copying: ".$row->file);
     my $file = $row->identifier.".netkan";
@@ -166,30 +167,65 @@ method inflate_random($number = 2) {
     push(@files, $output);
   }
 
-  $self->_inflator->inflate(\@files);
- 
-  # TODO: Once commited + pushed it will cause the mirror process to take place.
-  #       We should either mirror straight away and pull the commit process to be
-  #       done at the end
-  foreach my $row (@rows) {
+  my $inflator = App::KSP_CKAN::Crawler::InflateNetKAN->new(
+    config  => $self->config,
+  );
+  
+  # TODO: Pull out our status data so we can log dead metadata
+  my @successes = $inflator->inflate( 
+    files => \@files,
+    cache => $tmp."/cache",
+  );
+
+  my $mirror = App::KSP_CKAN::Crawler::MirrorCKAN->new(
+    config    => $self->config,
+    CKAN_meta => $self->_CKAN_meta,
+    cache     => $tmp."/cache",
+  );
+  foreach my $success (@successes) {
+    # Check we inflated the file without errors, then consume the metadata
+    $self->debug("Checking inflated file: ".$success);
+    my $index = first_index { $_->identifier eq basename($success,".netkan") } @rows;
+    my $row = $rows[$index];
     $self->debug("Consuming new metadata: ".$row->file);
-    my $ckan = App::KSP_CKAN::Metadata::Ckan->new( file => $path."/".$row->identifier."/".$row->file );
+    my $out_path = $path."/".$row->identifier;
+    my $ckan = App::KSP_CKAN::Metadata::Ckan->new( file => $out_path."/".$row->file );
     $row->update( {
       last_checked => \'NOW()',
       download_sha1 => $ckan->download_sha1,
     } );
+   
+    # A lot of metadata was handcrafted, potentially with incorrect names.
+    # lets remove the files we replace.
+    my $original_file = $out_path."/".$row->file;
+    my $new_file = $self->_check_output( $out_path, $original_file );
+    if ( $row->file ne $new_file ) {
+      $self->info($new_file." didn't match original filename of ".$row->file);
+      $self->info("Removing '".$row->file."' from metadata");
+      unlink($original_file) unless $self->is_debug;
+      $self->_CKAN_meta->commit(
+        file    => $original_file,
+        message => "'".$row->file."' replaced by '".$new_file,
+      ) unless $self->is_debug;
+    } else {
+      $self->debug($new_file." matched original filename of ".$row->file);
+    }
+
+    # Mirror this puppy!
+    $mirror->mirror($out_path."/".$new_file);
   }
 
-  remove_tree($tmp);
-  $self->_inflating(0);
+  $self->_inflator->push_ckan_meta unless $self->is_debug;
+  $self->debug("Temp Path: ".$tmp);
+  remove_tree($tmp) if ! $self->is_debug;
   return;
 }
 
 method run {
   # TODO: Ponder our timings
-  my $update_ckan_meta  = AE::timer 0,    3600, sub { $self->update_ckan_meta; };
-  my $check_existing    = AE::timer 15,   3600, sub { $self->check_existing; };
-  my $update_random     = AE::timer 30,   120,  sub { $self->update_random_ckans; };
+  #my $update_ckan_meta  = AE::timer 0,    3600, sub { $self->update_ckan_meta; };
+  #my $check_existing    = AE::timer 15,   3600, sub { $self->check_existing; };
+  #my $update_random     = AE::timer 30,   120,  sub { $self->update_random_ckans; };
 
   my $check_fork = AnyEvent::ForkManager->new(
       max_workers => 1,
@@ -197,28 +233,28 @@ method run {
       on_finish   => sub { $self->_mirror_check(0) },
   );
 
-  my $check_random = AE::timer 30, 120, sub { 
-    if ($self->_mirror_check) {
-       $self->debug("Skipping inflation this round");
-       return;
-    }
+  #my $check_random = AE::timer 30, 120, sub { 
+  #  if ($self->_mirror_check) {
+  #     $self->debug("Skipping mirror check this round");
+  #     return;
+  #  }
 
-    $check_fork->start(
-      cb => sub {
-        $self->check_random_mirrored; 
-      },
-    );
-  };
+  #  $check_fork->start(
+  #    cb => sub {
+  #      $self->check_random_mirrored; 
+  #    },
+  #  );
+  #};
 
   my $inflate_fork = AnyEvent::ForkManager->new(
       max_workers => 1,
-      on_start    => sub { $self->_mirror_check(1) },
-      on_finish   => sub { $self->_mirror_check(0) },
+      on_start    => sub { $self->_inflating(1) },
+      on_finish   => sub { $self->_inflating(0) },
   );
 
-  my $inflate_random =  AE::timer 60, 120, sub {
+  my $inflate_random =  AE::timer 0, 120, sub {
     if ($self->_inflating) {
-       $self->debug("Skipping mirror check this round");
+       $self->debug("Skipping inflation this round");
        return;
     }
